refactor: registry document class
- 등기부등본 데이터 클래스 리팩토링
