+++ node-connection-backend/src/main/java/node/connection/data/BuildingDescription.java
@@ -1,0 +1,24 @@
+package node.connection.data;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class BuildingDescription {
+    @JsonProperty("displayNumber")
+    private String displayNumber;
+
+    @JsonProperty("receiptDate")
+    private String receiptDate;
+
+    @JsonProperty("locationNumber")
+    private String locationNumber;
+
+    @JsonProperty("buildingDetails")
+    private String buildingDetails;
+
+    @JsonProperty("registrationCause")
+    private String registrationCause;
+}
