+++ node-connection-backend/src/main/java/node/connection/data/IssuanceData.java
@@ -1,0 +1,13 @@
+package node.connection.data;
+
+import java.time.LocalDateTime;
+
+public record IssuanceData(
+        String txId,
+        String issuerName,
+        String issuerDataHash,
+        RegistryDocument registryDocument,
+        LocalDateTime issuanceDate,
+        LocalDateTime expirationDate
+) {
+}
