+++ node-connection-backend/src/main/java/node/connection/data/IssuerData.java
@@ -1,0 +1,9 @@
+package node.connection.data;
+
+public record IssuerData(
+        String id,
+        String name,
+        String phoneNumber,
+        String email
+) {
+}
