+++ node-connection-backend/src/main/java/node/connection/data/RegistryDocument.java
@@ -1,0 +1,26 @@
+package node.connection.data;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+@Builder
+public class RegistryDocument {
+    @JsonProperty("id")
+    private String id;
+
+    @JsonProperty("titleSection")
+    private TitleSection titleSection;
+
+    @JsonProperty("exclusivePartDescription")
+    private ExclusivePartDescription exclusivePartDescription;
+
+    @JsonProperty("firstSection")
+    private List<FirstSection> firstSection;
+
+    @JsonProperty("secondSection")
+    private List<SecondSection> secondSection;
+}
