+++ node-connection-backend/src/main/java/node/connection/data/SecondSection.java
@@ -1,0 +1,24 @@
+package node.connection.data;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class SecondSection {
+    @JsonProperty("rankNumber")
+    private String rankNumber;
+
+    @JsonProperty("registrationPurpose")
+    private String registrationPurpose;
+
+    @JsonProperty("receiptDate")
+    private String receiptDate;
+
+    @JsonProperty("registrationCause")
+    private String registrationCause;
+
+    @JsonProperty("holderAndAdditionalInfo")
+    private String holderAndAdditionalInfo;
+}
