--- node-connection-backend/src/main/java/node/connection/dto/registry/request/AddBuildingDescriptionToTitleSection.java
@@ -1,9 +1,0 @@
-package node.connection.dto.registry.request;
-
-import node.connection.dto.registry.BuildingDescriptionDto;
-
-public record AddBuildingDescriptionToTitleSection(
-        String documentId,
-        BuildingDescriptionDto buildingDescription
-) {
-}
