--- node-connection-backend/src/main/java/node/connection/dto/registry/request/AddBuildingPartDescriptionToExclusivePart.java
@@ -1,9 +1,0 @@
-package node.connection.dto.registry.request;
-
-import node.connection.dto.registry.BuildingPartDescriptionDto;
-
-public record AddBuildingPartDescriptionToExclusivePart(
-        String documentId,
-        BuildingPartDescriptionDto buildingPartDescription
-) {
-}
