--- node-connection-backend/src/main/java/node/connection/dto/registry/request/AddFirstSectionEntry.java
@@ -1,9 +1,0 @@
-package node.connection.dto.registry.request;
-
-import node.connection.dto.registry.FirstSectionDto;
-
-public record AddFirstSectionEntry(
-        String documentId,
-        FirstSectionDto firstSection
-) {
-}
