--- node-connection-backend/src/main/java/node/connection/dto/registry/request/AddLandDescriptionToTitleSection.java
@@ -1,9 +1,0 @@
-package node.connection.dto.registry.request;
-
-import node.connection.dto.registry.LandDescriptionDto;
-
-public record AddLandDescriptionToTitleSection(
-        String documentId,
-        LandDescriptionDto landDescription
-) {
-}
