--- node-connection-backend/src/main/java/node/connection/dto/registry/request/AddLandRightDescriptionToExclusivePart.java
@@ -1,9 +1,0 @@
-package node.connection.dto.registry.request;
-
-import node.connection.dto.registry.LandRightDescriptionDto;
-
-public record AddLandRightDescriptionToExclusivePart(
-        String documentId,
-        LandRightDescriptionDto landRightDescription
-) {
-}
