--- node-connection-backend/src/main/java/node/connection/dto/registry/request/AddSecondSectionEntry.java
@@ -1,9 +1,0 @@
-package node.connection.dto.registry.request;
-
-import node.connection.dto.registry.SecondSectionDto;
-
-public record AddSecondSectionEntry(
-        String documentId,
-        SecondSectionDto secondSection
-) {
-}
