--- node-connection-backend/src/main/java/node/connection/dto/registry/request/FindUserRegistryDocumentsRequest.java
@@ -1,7 +1,0 @@
-package node.connection.dto.registry.request;
-
-public record FindUserRegistryDocumentsRequest(
-        String phoneNumber,
-        String password
-) {
-}
