--- node-connection-backend/src/main/java/node/connection/dto/registry/request/RegistryCreateRequest.java
@@ -1,13 +1,0 @@
-package node.connection.dto.registry.request;
-
-import node.connection.dto.registry.RegistryDocumentDto;
-
-public record RegistryCreateRequest(
-        String phoneNumber,
-        String name,
-        String birthDate,
-        String email,
-        String password,
-        RegistryDocumentDto document
-) {
-}
